Major changes: enteryActivity
general changes:
- converted IDs to long
-Add constants: for transaction types, for reason types
-removed unneeded code from classes

+++ branches/MintTrackV2/MintTrack/src/com/ponyinc/minttrack/Accounts.java
@@ -39,7 +39,7 @@ public class Accounts {
 	/** Method used to query the accounts table based on an accounts ID
 	*	@para _id Account ID that you wish to query
 	*/
-	public Cursor getAccount(int _id) {
+	public Cursor getAccount(long _id) {
 		final String[] FROM = { _ID, ACCOUNT_NAME, ACCOUNT_TOTAL,
 				ACCOUNT_ACTIVE, };
 		final String ORDER_BY = _ID + " DESC";
@@ -81,7 +81,7 @@ public class Accounts {
 	/** Method is used to deactivate an account
 	*	@param acc_id ID of the account which you wish to disable
 	*/
-	public void DeactivateAccount(int acc_id)
+	public void DeactivateAccount(long acc_id)
 	// set account to inactive
 	{
 		SQLiteDatabase db = MintLink.getWritableDatabase();
@@ -93,7 +93,7 @@ public class Accounts {
 	*	@param acc_id ID of the account you want to modify
 	*	@param strName Name that will replace the account
 	*/
-	public void EditAccountName(int acc_id, String strName) {
+	public void EditAccountName(long acc_id, String strName) {
 		SQLiteDatabase db = MintLink.getWritableDatabase();
 		ContentValues values = new ContentValues();
 		values.put(ACCOUNT_NAME, strName);
@@ -103,7 +103,7 @@ public class Accounts {
 	*	@param acc_id ID of the account you want to modify
 	*	@param total New amount to be set as the balance of the account
 	*/
-	public void EditAccountTotal(int acc_id, double total) {
+	public void EditAccountTotal(long acc_id, double total) {
 		SQLiteDatabase db = MintLink.getWritableDatabase();
 		ContentValues values = new ContentValues();
 		values.put(ACCOUNT_TOTAL, total);
@@ -112,7 +112,7 @@ public class Accounts {
 	/** Method is used to reactive an inactive account that already exists
 	*	@param acc_id ID of account which you want to activate
 	*/
-	public void ReactivateAccount(int acc_id)
+	public void ReactivateAccount(long acc_id)
 	// set account to inactive
 	{
 		SQLiteDatabase db = MintLink.getWritableDatabase();
